backend/requirements.txt
agno
fastapi
uvicorn[standard]
openai
python-dotenv
googlemaps
